pyproject.toml
[project]
name = "scenario-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
